requirements.txt
flask
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
openai
chromadb
python-dotenv
gunicorn
waitress